auth(fix): address PR feedback and CI failures
