+++ server/db/schema/index.ts
@@ -23,6 +23,7 @@ export * from './roles.ts';
 export * from './sales.ts';
 export * from './settings.ts';
 export * from './sso.ts';
+export * from './ssoProviders.ts';
 export * from './supplierInvoices.ts';
 export * from './supplierOrderVersions.ts';
 export * from './supplierQuoteAttachments.ts';
